pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "openai>=1.95.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "urllib3>=2.5.0",
]